Logging: Can now specify log_folder in Setup_State
This is useful to specify the log_folder before the initial blocks get
logged. Otherwise these initial blocks always get logged in the log_folder
from input.cfg.
When orchestrating calculations with the python API this can be annoying,
because setting the log folder with the API function only takes effect *after*
the state creation, which means that some initial messages get logged in
a different location.
Also enabled setting the log folder via the CLI arguments of main.