API(EMA): get_modes and get_eigenvalues
Wrote C- and python-API functions to retrieve the eigenvalues and
eigenmodes after an EMA caluclation. The C-API functions work by copying
into a buffer that has to be allocated by the caller.
TODO: check if this implementation leaks memory ... I am unsure if the
allocation of the buffer is handled correctly on the ctypes side.