feature: semi-classical Monte Carlo sampling
This implements a semi-classical Monte Carlo trial move. Semi-classical
here means that there is an integer spin quantum number `spin_qn` that
restricts the component along the local effective field to be on one of
the `spin_qn + 1` level sets in [-1, 1]. That means that `spin_qn`
relates to the conventional spin quantum number s by `s = spin_qn / 2`.